+++ Restaurants/Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantByIdCommandHandler.cs
@@ -1,11 +1,15 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
+using Restaurants.Infrastructure.Authorization;
+using Restaurants.Infrastructure.Authorization.Services;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantByIdCommandHandler(
         ILogger<DeleteRestaurantByIdCommandHandler> logger,
+        IRestaurantAuthorizeService restaurantAuthorizeService,
         IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantByIdCommand, bool>
     {
         public async Task<bool> Handle(DeleteRestaurantByIdCommand request, CancellationToken cancellationToken)
@@ -16,6 +20,11 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
             if (restaurant != null) 
             {
                 logger.LogInformation($"Deleting restaurant with id: {request.Id} form db");
+                if(!restaurantAuthorizeService.Authorize(restaurant, ResourceOperation.Delete))
+                {
+                    throw new ForbidException();
+                }
+
                 isDeleted = await restaurantsRepository.DeleteRestaurantById(restaurant);
             }
